moved url & aem site grabber to playwright config
