+++ theme/playwright.config.js
@@ -1,8 +1,28 @@
+const fs = require("fs");
+
+const getLiveServerDetails = () => {
+  const output = {};
+  const text = fs.readFileSync("../theme/.env", "utf8");
+  const lines = text.split(/\r\n|\n/);
+  const siteIndex = lines.findIndex((string) => string.startsWith("AEM_SITE="));
+  const portIndex = lines.findIndex((string) =>
+    string.startsWith("AEM_PROXY_PORT=")
+  );
+
+  output.aemSite = lines[siteIndex].split("=")[1];
+  output.aemPort = Number(lines[portIndex].split("=")[1]);
+
+  return output;
+};
+const liveServerDetails = getLiveServerDetails();
+
 const config = {
+  aemSite: liveServerDetails.aemSite,
+
   timeout: 30000, // Timeout is shared between all tests.
   webServer: {
     command: "npm run live",
-    port: 7000,
+    port: liveServerDetails.aemPort,
   },
   use: {
     headless: true,
